% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{adjacency_matrix.read_table}
\alias{adjacency_matrix.read_table}
\title{Create edge matrix for read graph}
\usage{
adjacency_matrix.read_table(df, sparse = T)
}
\arguments{
\item{df}{read_table object}

\item{sparse}{If TRUE, remove redundant edges}
}
\value{
a square matrix with dimension equal to the number
of read groups in the read graph
}
\description{
Create edge matrix for read graph
}
