% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_engine.R
\name{AL.engine}
\alias{AL.engine}
\title{Compute the augmented lagrangian}
\usage{
AL.engine(y, P, M, rho, mu, kk, pi)
}
\description{
Compute the augmented lagrangian
}
\details{
This function is not used in the optimization and is only for debugging
purposes
}

