% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{all_pos.read_table}
\alias{all_pos.read_table}
\title{Return the positions covered by each read group in a read table}
\usage{
all_pos.read_table(df)
}
\arguments{
\item{df}{read_table object}
}
\value{
a list of numeric vectors containing read positions
}
\description{
Return the positions covered by each read group in a read table
}

