% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{BAM_file_position_pair_frequencies.error}
\alias{BAM_file_position_pair_frequencies.error}
\title{Extract the position pair frequencies from read table generated from BAM file}
\usage{
BAM_file_position_pair_frequencies.error(out_dir, pair_df)
}
\arguments{
\item{out_dir}{directory in which to find `pair_comparison_read_table.csv`}

\item{pair_df}{data.frame produced by position_pairs.error giving
position pairs to consider}
}
\value{
a data.frame with columns pos1, pos2, nuc1, nuc2, nuc, freq
where nuc1, nuc2 are the nucleotides at pos1 and pos2, nuc is nuc1+nuc2,
and freq is the frequency
}
\description{
Extract the position pair frequencies from read table generated from BAM file
}

