% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{BAM_pileup}
\alias{BAM_pileup}
\title{Calculate pileup for each position of reference from BAM file}
\usage{
BAM_pileup(bam_file, bai_file = NULL, max_depth = 2000,
  min_base_quality = 30, min_mapq = 30, strand = NULL)
}
\arguments{
\item{bam_file}{bam file}

\item{bai_file}{If not passed, then assumed to be same as bam_file
with .bai appendix.}

\item{max_depth}{maximum number of reads from which to form the pileup.
This parameter is passed to RSamTools PileupParam function.}

\item{min_base_quality}{Passed to PileupParam}

\item{min_mapq}{Passed to PileupParam}

\item{strand}{If BAM file is paired end, which strand to use: "plus" or "minus".  Should
be set to NULL if all reads to be used or if BAM file is not paired.}

\item{start_pos}{Start position of pileup relative to reference.}

\item{end_pos}{End position of pileup relative to reference.}
}
\value{
a VSeqTools_pileup object which inherits data.frame
and has columns: pos, A, C, G, T, i, d, Coverage.
pos gives the position on the reference.  Coverage is total reads
at that position.  The rest of the columns give read freq at that position
with the particular nt value.
}
\description{
Using the Rsamtools pileup function, the counts of A,C,G,T,i,d
are determined at each position on reference
}

