% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{bam_to_variant_calls.pipeline}
\alias{bam_to_variant_calls.pipeline}
\title{Given a bam file, create variant call file}
\usage{
bam_to_variant_calls.pipeline(bam_file, out_dir, start_pos = NULL,
  end_pos = NULL, sig = 0.01, heavy_tail = T)
}
\arguments{
\item{bam_file}{path to bam file}

\item{out_dir}{output directory, will be created if needed.  variant call file
will be placed in the output directory}

\item{start_pos}{position on reference at which to start looking for variant
calls}

\item{end_pos}{position on reference at which to end looking for variant calls}

\item{sig}{significance level at which variants will be called.  An automatic
bonferroni correction is applied to account for the number of positions in the
reference.}

\item{heavy_tail}{If T, then a betabinomial is used to call errors following Gerstung et al
2011. If F, then a poisson is used to call variants following Wang et al 2007}
}
\value{
path to variant call file
}
\description{
Given a bam file, create variant call file
}
\details{
variant calls are given by a data.frame with 6 columns
named c("pos", "A", "C", "G", "T", "-") pos gives the position at which a
true variant exists.  The other columns have logical entries,
with TRUE meaning that a true variant exists with the corresponding nucleotide.
}
\references{
Wang,C. et al. (2007) Characterization of mutation
spectra with ultra-deep pyrosequencing: application to HIV-1 drug
resistance. Genome Res., 17, 1195–1201,

Gerstung et al. (2011) Reliable detection of subclonal
single-nucleotide variants in tumour cell populations.  Nature
Communications.
}

