% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_analyze_solutions.R
\name{best_K.RegressHaploSolutions}
\alias{best_K.RegressHaploSolutions}
\title{Determines the best fit solution}
\usage{
best_K.RegressHaploSolutions(rhs)
}
\arguments{
\item{rhs}{A RegressHaploSolutions object}
}
\description{
Determines the best fit solution
}
\details{
K is given by the number of non-zero entries in pi returned
by RegressHaplo
}

