% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{calculate_position_pair_error.error}
\alias{calculate_position_pair_error.error}
\title{Calculate error for each position pair}
\usage{
calculate_position_pair_error.error(df1, df2)
}
\arguments{
\item{df1}{A data.frame containing pos1, pos2, nuc1, nuc2, nuc (pasted nuc1,nuc2),
freq (produced by BAM_file_position_pair_frequencies.error or
reconstructed_position_pair_frequencies.error)}
}
\value{
a data.frame containing pos1, pos2, error where error is sum of absolute
frequency differences at each position over true/predicted nucleotide pairs.
}
\description{
Calculate error for each position pair
}

