% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{clean.read_table}
\alias{clean.read_table}
\title{Clean read table}
\usage{
clean.read_table(df, min_count = 10, remove_outside_reads = T,
  remove_empty_cols = T, remove_non_variant_pos = F, remove_deletions = F,
  remove_partial_cover_reads = F)
}
\arguments{
\item{df}{read table}

\item{min_count}{remove rows with count below}

\item{remove_outside_reads}{remove rows with all NA}

\item{remove_empty_cols}{remove cols with all NA}

\item{remove_non_variant_pos}{remove cols with only 1 value other than NA}

\item{remove_deletions}{replace "-" with NA}

\item{remove_partial_cover_reads}{remove rows with one or more NA}
}
\description{
Clean read table
}

