% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplo.R
\name{compare.Haplo}
\alias{compare.Haplo}
\title{Compare frequencies and haplotypes of H1 using
the frequencies and haplotype of H2 as a base}
\usage{
\method{compare}{Haplo}(H1, H2)
}
\description{
Calculate the distance from the H1 haplotype to the
closest H2 haplotype.  Calculate the difference between
the frequencies of H1 and the frequencies of H2
}
