% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_util.R
\name{compute_solution.RegressHaplo}
\alias{compute_solution.RegressHaplo}
\title{Fit regression for a given rho}
\usage{
compute_solution.RegressHaplo(df, h_full, rho, verbose = T)
}
\arguments{
\item{df}{read table}

\item{h_full}{haplotype matrix}

\item{rho}{rho to be used in regression}
}
\value{
A list with entries pi and fit.  pi gives the frequencies
estimated for h_full and fit gives a fit value.
}
\description{
Fit regression for a given rho
}
