% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{consensus.read_table}
\alias{consensus.read_table}
\title{Return the consensus sequence at each position of the read table}
\usage{
consensus.read_table(df)
}
\description{
Return the consensus sequence at each position of the read table
}
