% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{consistent_haplotypes.read_table}
\alias{consistent_haplotypes.read_table}
\title{Split read table into loci with no spanning single end reads, create
consistent haplotypes for each locus, and then combine to create full
haploptypes}
\usage{
consistent_haplotypes.read_table(df_in, rm.na = F,
  max_num_haplotypes = 20000)
}
\description{
Split read table into loci with no spanning single end reads, create
consistent haplotypes for each locus, and then combine to create full
haploptypes
}
