% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{consistent_haplotypes_across_loci.read_table}
\alias{consistent_haplotypes_across_loci.read_table}
\title{Split read table into loci with no spanning reads, create consistent
haplotypes for each locus, and then combine to create full haplotypes}
\usage{
consistent_haplotypes_across_loci.read_table(
  df_in,
  min_cover = 500,
  max_num_haplotypes = 20000
)
}
\description{
Split read table into loci with no spanning reads, create consistent
haplotypes for each locus, and then combine to create full haplotypes
}
