% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{coverage.read_table}
\alias{coverage.read_table}
\title{Return coverage at each position in read table}
\usage{
coverage.read_table(df)
}
\value{
A numeric vector
}
\description{
Return coverage at each position in read table
}

