% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{create_position_pair_comparison_read_table.error}
\alias{create_position_pair_comparison_read_table.error}
\title{Create a read table for analyzing frequencies of positions pairs.  This
is a pre-analysis step.}
\usage{
create_position_pair_comparison_read_table.error(bam_file, variable_calls,
  out_dir)
}
\arguments{
\item{bam_file}{path to BAM file}

\item{out_dir}{directory in which to write the output read table as
`position_pair_read_table.csv`}

\item{variant_calls}{a variant call object as outputed by variant_calls()}
}
\value{
path to created file
}
\description{
Create a read table for analyzing frequencies of positions pairs.  This
is a pre-analysis step.
}
