% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{create_refspace_seq}
\alias{create_refspace_seq}
\title{Converts reads to a sequence in the reference space}
\usage{
create_refspace_seq(seq, cigar)
}
\arguments{
\item{seq}{Read sequences}

\item{cigar}{Read cigars}
}
\value{
A vector of sequences in the reference space.  Insertions are ignored since they
are not in the reference space
}
\description{
Given a vector of read sequences and their corresponding cigars, returns the sequences
in the reference space by removing soft clippings, removing insertions (relative to
the reference), and inserting
"-" for deletions (relative to the reference)
}
