% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{divergence_variant_calls}
\alias{divergence_variant_calls}
\title{Calls nt positions at which a BAM file differs from a base BAM file}
\usage{
divergence_variant_calls(BAMfile, b_BAMfile, min_divergence_freq = 0.02)
}
\arguments{
\item{BAMfile}{BAM file}

\item{b_BAMfile}{base BAM file to which BAMfile is to be compared}

\item{min_divergenece_freq}{minimum frequency for which divergence is called}
}
\value{
A numeric vector of called nt positions.
}
\description{
The pileup of the first BAM file is compared to the consensus of the
second BAM file and positions at which there are disagreements are called
}
\details{
insertions are ignored in calling divergent positions
}
