% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{end_pos.read_table}
\alias{end_pos.read_table}
\title{Return the end position of each read group in a read table}
\usage{
end_pos.read_table(df)
}
\arguments{
\item{df}{read_table object}
}
\value{
a numeric vector of start positions
}
\description{
Return the end position of each read group in a read table
}

