% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{error_filter.read_table}
\alias{error_filter.read_table}
\title{Remove reads that are below error noise.}
\usage{
error_filter.read_table(df, error_freq, sig)
}
\arguments{
\item{df}{read table}

\item{error_freq}{per position error rate}

\item{sig}{significance level at which to filter out errors.}
}
\value{
a read table with errors that are below noise threshold
removed
}
\description{
Remove reads that are below error noise.
}
