% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{filter_true_variants.read_table}
\alias{filter_true_variants.read_table}
\title{Filter read table for variant calls}
\usage{
filter_true_variants.read_table(df, variant_calls, consensus_values)
}
\arguments{
\item{df}{read_table}

\item{variant_calls}{variant calls data.frame, see read_table() help}

\item{consensus_values}{consensus nucleotides at variable positions}
}
\value{
a data.frame in which errors (i.e. not called varaints) are replaced
by the consensus value.
}
\description{
Filter read table for variant calls
}
\details{
nrow of variant calls must equal length of conseunsus values
}

