% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{get_error_rate}
\alias{get_error_rate}
\title{Estimate error rate of pileup}
\usage{
get_error_rate(pu, split_by_nuc = F)
}
\arguments{
\item{pu}{a VSeqTools_pileup object returned by BAM_pileup}

\item{split_by_nuc}{If T, an error rate is returned for each nucleotide,
otherwise a single error rate is returned representing sum of all possible errors}
}
\value{
A vector of error rates.
}
\description{
Estimate error rate of pileup
}

