% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{get_position_pair_comparison_read_table.error}
\alias{get_position_pair_comparison_read_table.error}
\title{Get `position_pair_read_table.csv` from directory}
\usage{
get_position_pair_comparison_read_table.error(out_dir)
}
\description{
Get `position_pair_read_table.csv` from directory
}
