% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{get_solutions_haplotype_reconstruction.pipeline}
\alias{get_solutions_haplotype_reconstruction.pipeline}
\title{Haplotype reconstruction of given solution}
\usage{
get_solutions_haplotype_reconstruction.pipeline(out_dir, i)
}
\arguments{
\item{out_dir, }{RegressHaplo pipeline directory}

\item{i}{solution number}
}
\value{
a data.frame with columns (haplotype) haplotypes reconstructed, (freq)
frequencies of haplotype, and the haplotype number from the list of global
haplotypes in h.csv (haplotype_number)
}
\description{
Haplotype reconstruction of given solution
}

