% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{get_solutions_summary.pipeline}
\alias{get_solutions_summary.pipeline}
\title{Summary of regression fits over all starting points}
\usage{
get_solutions_summary.pipeline(out_dir)
}
\arguments{
\item{out_dir}{RegressHaplo pipeline directory}
}
\value{
A data.frame with columns (rho) rho value for fitting,
(K) number of haplotypes reconstruted,
(fit) fit of haplotype reconstruction, (solution_number).  Each
row of the data.frame corresponds to a solution of the optimization
for a given starting point and rho value.
}
\description{
Summary of regression fits over all starting points
}
