% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{get_variant_call_pos}
\alias{get_variant_call_pos}
\title{Get variant call positions from a VSeqTools_variant_call
object}
\usage{
get_variant_call_pos(vc)
}
\arguments{
\item{vc}{VSeqTools_variant_call object}
}
\value{
a numeric vector giving positions relative to reference at which
variants are called.
}
\description{
Get variant call positions from a VSeqTools_variant_call
object
}
