% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplo.R
\name{Haplo}
\alias{Haplo}
\title{Haplo object construct}
\usage{
Haplo(h, pi, count = NA)
}
\arguments{
\item{h}{A haplotype matrix (matrix of characters with rows giving haplotypes)}

\item{pi}{A numeric vector giving frequencies of haplotypes}

\item{count}{The total coverage, may be NA if this is not well-defined}
}
\value{
A Haplo object
}
\description{
A Haplo object contains a collection of haplotypes,
their frequency, and total count
}
\details{
A Haplo object is a list with elements h, pi, count, h_seq.
h is a character matrix with rows giving haplotypes, pi is a numeric
vector of frequencies, count is not currently used and h_seq is
a characer vector of haplotypes (rows of h pasted).
}
