% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFit.R
\name{haplotype_match.readFit}
\alias{haplotype_match.readFit}
\title{Determine haplotypes that match a nucleotide pattern}
\usage{
haplotype_match.readFit(nucs, h)
}
\arguments{
\item{nucs}{A character vector equal to the number of positions
in the read table.  Positions at which nuceotides are not specified
have value "+".}

\item{df}{read table}

\item{haplotype}{matrix}
}
\value{
A numeric vector of 1's and 0's
}
\description{
Given a nucleotide pattern covering some portion of
read table positions, return a vector of 1's and 0's
representing haplotypes that match the patterh
}
