% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{haplotypes_to_fasta.pipeline}
\alias{haplotypes_to_fasta.pipeline}
\title{Given RegressHaplo haplotypes file, produce fasta file showing
haplotypes over the full reference with frequencies as part of
sequence names}
\usage{
haplotypes_to_fasta.pipeline(bam_file, out_dir)
}
\arguments{
\item{bam_file}{bam file containing reads for haplotype reconstruction}

\item{out_dir}{output directory for haplotype file and assumed
directory of solutions file}
}
\value{
path to fasta file
}
\description{
Given RegressHaplo haplotypes file, produce fasta file showing
haplotypes over the full reference with frequencies as part of
sequence names
}

