% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{loci_to_haplotypes.pipeline}
\alias{loci_to_haplotypes.pipeline}
\title{Given loci, create global haplotypes}
\usage{
loci_to_haplotypes.pipeline(out_dir, max_num_haplotypes = 1000)
}
\arguments{
\item{out_dir}{output directory for parameter files and assumed
directory of read table}
}
\description{
Given loci, create global haplotypes
}

