% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_engine.R
\name{max_Hessian_eigenvalue.engine}
\alias{max_Hessian_eigenvalue.engine}
\title{Compute maximum eigenvalue of Hessian to scale gradient}
\usage{
max_Hessian_eigenvalue.engine(QQ)
}
\description{
Compute maximum eigenvalue of Hessian to scale gradient
}

