% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{merge_haplotypes.read_table}
\alias{merge_haplotypes.read_table}
\title{A helper function that create haplotype permutations from a list
of haplotype matrices}
\usage{
merge_haplotypes.read_table(con_haps)
}
\arguments{
\item{con_haps}{A list of haplotype matrices}
}
\description{
A helper function that create haplotype permutations from a list
of haplotype matrices
}
