% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{nucs_at_pos.read_table}
\alias{nucs_at_pos.read_table}
\title{Return nucleotide counts at each position}
\usage{
nucs_at_pos.read_table(df)
}
\arguments{
\item{df}{read_table}
}
\value{
A matrix with 5 rows and a column for each position in df
}
\description{
Return nucleotide counts at each position
}

