% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_create_solutions.R
\name{parameters.RegressHaplo}
\alias{parameters.RegressHaplo}
\title{Generate the y vector and P matrix of the RegressHaplo optimization and
the associated haplotyes}
\usage{
\method{parameters}{RegressHaplo}(df, max_global_dim = 1200,
  max_local_dim = 1200, min_cover = 500)
}
\value{
a list with entries y (numeric), P(numeric matrix) and h
(a character matrix).
}
\description{
Generates y and P using a call to filter_and_optimize.RegressHaplo and
also returns a matrix of consistent haploytpes
}

