% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{parameters_to_solutions.pipeline}
\alias{parameters_to_solutions.pipeline}
\title{Given a parameter files, run RegressHaplo to produce multiple solutions}
\usage{
parameters_to_solutions.pipeline(out_dir, num_trials = 700,
  rho_vals = c(0.1, 1, 5, 10, 20))
}
\arguments{
\item{out_dir}{output directory for parameter files and assumed
directory of parameter files}

\item{num_trials}{number of starting points to try in attempting to find
a global min for each rho values}

\item{rho_values}{rho values to use in finding global min.}
}
\value{
path to haplotype file
}
\description{
Given a parameter files, run RegressHaplo to produce multiple solutions
}
\details{
The parameter files assumed to exist are y.csv, P.csv, h.csv
}
