% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_analyze_solutions.R
\name{parse_solutions.RegressHaploSolutions}
\alias{parse_solutions.RegressHaploSolutions}
\title{Parse solutions matrix}
\usage{
parse_solutions.RegressHaploSolutions(solutions)
}
\arguments{
\item{filename}{CSV filename}

\item{animal}{animal to which this CSV file is associated}

\item{wekk}{week to which this CSV file is associated}
}
\value{
A data.frame with the first four rows fit, rho, K, kk
followed by columns of dimension p, giving the optimal result.
}
\description{
Parse solutions matrix
}
\details{
A solutions matrix is a matrix with each column
corresponding to a solution.
Within each column, the first four rows are
fit, rho, K, and kk followed the computed solution.
}

