% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{paths_exceed_limit.read_table}
\alias{paths_exceed_limit.read_table}
\title{Quickly determine the number of paths in an adjacency matrix exceed
a lower limit}
\usage{
paths_exceed_limit.read_table(m, paths_cutoff)
}
\arguments{
\item{m}{adjacency matrix}

\item{paths_cutoff}{threshold for number of paths}
}
\value{
NA if number of paths exceeds num_paths, otherwise
returns the number of paths.
}
\description{
Quickly determine the number of paths in an adjacency matrix exceed
a lower limit
}
\details{
number of paths are determined by dynamic programming algorith
that is O(V+E).  The quick part is that once the number of paths
exceeds num_paths, the algorithm exits.   This is needed when the number
of edges is very large and we would waste time computing the total
number of paths.
}

