% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{paths_to_haplotypes.read_table}
\alias{paths_to_haplotypes.read_table}
\title{A helper function that converts read graph paths to haplotype matrices}
\usage{
paths_to_haplotypes.read_table(paths, df)
}
\arguments{
\item{paths}{a list of paths generated by igraph function all_simple_paths}

\item{df}{the read table from which paths was built}
}
\value{
A haplotype matrix
}
\description{
A helper function that converts read graph paths to haplotype matrices
}
