% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_util.R
\name{penalized_regression.RegressHaplo}
\alias{penalized_regression.RegressHaplo}
\title{Solves min_pi |y - P*pi|^2 + rho*pi^T*M*pi}
\usage{
penalized_regression.RegressHaplo(y, P, pi = NULL, rho, kk = 2,
  verbose = T)
}
\description{
Solves min_pi |y - P*pi|^2 + rho*pi^T*M*pi
}

