% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_util.R
\name{penalized_regression_parameters.RegressHaplo}
\alias{penalized_regression_parameters.RegressHaplo}
\title{Returns the matrices and vectors associated with the regression
optimization}
\usage{
penalized_regression_parameters.RegressHaplo(df, h, position_fit = F)
}
\arguments{
\item{df}{read table}

\item{h}{haplotype matrix}
}
\value{
the matrix P and the vector y as a list.  M is implicit
and so not computed.
}
\description{
Returns the matrices and vectors associated with the regression
optimization
}
\details{
The regression solves
min |y - P*pi|^2 + rho pi^T*M*pi
subject to:  sum pi_i = 1, pi_i ge 0.
}
