% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFit.R
\name{plot.readFit}
\alias{plot.readFit}
\title{Visualize a readfit object}
\usage{
plot.readFit(rf, outfile = NULL)
}
\arguments{
\item{rf}{a readFit object}

\item{outfile}{If not null, figure is written to outfile.}
}
\description{
Visualize a readfit object
}

