% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{plot_pair_position_error.error}
\alias{plot_pair_position_error.error}
\title{Plot the position errors}
\usage{
plot_pair_position_error.error(df)
}
\arguments{
\item{df}{A data.frame produced by calculate_position_pair_error}

\item{boxplot}{If T, show a boxplot, otherwise barplot}
}
\description{
Plot the position errors
}

