% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{plot_single_position.error}
\alias{plot_single_position.error}
\title{Plot the position errors}
\usage{
plot_single_position.error(df, boxplot = T)
}
\arguments{
\item{df}{data.frame with pos and error columns, both numeric}

\item{boxplot}{If T, show a boxplot that considers all errors collectivly,
otherwise show a barplot giving error at each position}
}
\description{
Plot the position errors
}

