% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{pos_names.read_table}
\alias{pos_names.read_table}
\title{Returns positions of read table as character vector}
\usage{
pos_names.read_table(df)
}
\value{
A character vector
}
\description{
Returns positions of read table as character vector
}

