% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFit.R
\name{position_fit.readFit}
\alias{position_fit.readFit}
\title{Return a readFit object base on nucleotide positions}
\usage{
position_fit.readFit(df, h, pi = NULL)
}
\description{
Return a readFit object base on nucleotide positions
}
