% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{position_pairs.error}
\alias{position_pairs.error}
\title{Get position pairs with threshold coverage}
\usage{
position_pairs.error(out_dir, max_pair_distance = 200, minimum_coverage = 1000)
}
\arguments{
\item{out_dir}{directory in which to find pair_comparison_read_table}

\item{max_pair_distance}{maximum distance to allow between pairs}

\item{minimum_coverage}{minimum numbers of reads covering a pair}
}
\value{
a data.frame with columns pos1, pos2 giving the positions of the pairs
}
\description{
Get position pairs with threshold coverage
}
