% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{read_BAM_pileup}
\alias{read_BAM_pileup}
\title{Read a VSeqTools_pileup object}
\usage{
read_BAM_pileup(filename)
}
\arguments{
\item{filename}{A csv file}
}
\value{
a VSeqTools_pileup object
}
\description{
Read a VSeqTools_pileup object
}
