% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{read_table_to_loci.pipeline}
\alias{read_table_to_loci.pipeline}
\title{Given a read table, split variable positions into loci}
\usage{
read_table_to_loci.pipeline(out_dir, max_num_haplotypes = 1200,
  min_cover = 500)
}
\arguments{
\item{out_dir}{output directory for parameter files and assumed
directory of read table}
}
\description{
Given a read table, split variable positions into loci
}
\details{
The read table is assumed to be in out_dir with filename read_table.csv.
}

