% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{reads_covering_haplotypes.read_table}
\alias{reads_covering_haplotypes.read_table}
\title{For each haplotype, return reads that match haplotype at covered positions}
\usage{
reads_covering_haplotypes.read_table(df, h)
}
\arguments{
\item{df}{read table}

\item{h}{haplotype matrix}
}
\value{
A R by K logical matrix where R is number of reads and K is
number of haplotypes.  A TRUE entry means the read covers the haplotype.
}
\description{
For each haplotype, return reads that match haplotype at covered positions
}
