% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{reconstructed_position_pair_frequencies.error}
\alias{reconstructed_position_pair_frequencies.error}
\title{Extract the position pair frequencies from haplotypes}
\usage{
reconstructed_position_pair_frequencies.error(haplotypes, freq, pair_df)
}
\arguments{
\item{haplotypes}{a character vector of haplotypes}

\item{freq}{a numeric vector of haplotype frequencies}

\item{pair_df}{data.frame produced by position_pairs.error giving
position pairs to consider}
}
\value{
a data.frame with columns pos1, pos2, nuc1, nuc2, nuc, freq
where nuc1, nuc2 are the nucleotides at pos1 and pos2, nuc is nuc1+nuc2,
and freq is the frequency
}
\description{
Extract the position pair frequencies from haplotypes
}
