% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_analyze_solutions.R
\name{RegressHaploSolutions}
\alias{RegressHaploSolutions}
\title{Loads a set of solutions created by solutions.RegressHaplo}
\usage{
RegressHaploSolutions(solutions, h, Igor = F)
}
\arguments{
\item{solutions}{Output from solutions.RegressHaplo}

\item{h}{Haplotypes used to construct the y and P matrix used
in generating solutions}

\item{Igor}{For now Igor's format and mine are slightly different!}
}
\value{
A list containing the entries df_stats, df_p,
read_df, founder_df, h_full, and results.
df_stats is a data.frame
containing the columns run_id,
rho, K, Igor_K, fit.  m_p is a matrix with rows p1, ..., pn;  see
\code{\link{parse_Igor_csv_to_df.IgorResults}} for further
details.  read_df and founder_df are the read_table for the
dataset and founder table for the animal, respectively.
h_full is the collection of hapltypes corresponding to the
entries of p.  results is a list, with an entry for each row
of results_df, i.e. each optimization.  Each entry of results
is a list containing h (haplotype matrix), pi (haplotype frequencies),
and fit.
}
\description{
Loads a set of solutions created by solutions.RegressHaplo
}

