% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{regroup.read_table}
\alias{regroup.read_table}
\title{Merges identical read over an edited read table}
\usage{
regroup.read_table(df)
}
\arguments{
\item{df}{a read_table object}
}
\value{
a read_table object
}
\description{
Merges identical read over an edited read table
}
\details{
After removing columns for positions that are
not of interest, a read table can contain multiple rows
corresponding to read groups that are identical at the remaining positions.
This functions joins those reads and returns a read_table
with unique read groups.
}

