% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{seq.read_table}
\alias{seq.read_table}
\title{Return the nucleotides in each read group of a read table}
\usage{
\method{seq}{read_table}(df)
}
\arguments{
\item{df}{read_table object}
}
\value{
a list of character vectors containing read nucleotides
}
\description{
Return the nucleotides in each read group of a read table
}
