% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplo.R
\name{set_pos_names.Haplo}
\alias{set_pos_names.Haplo}
\title{Set the position names for the characters (nucleotides) forming the haplotypes.}
\usage{
set_pos_names.Haplo(H, pos_names)
}
\arguments{
\item{H}{A Haplo object}

\item{pos_names}{A character vector giving position names}
}
\description{
Alters the colnames of the haplotype matrix
to reflect pos names.   This is meant for visual purposes.   The Haplo object never accesses
these names.
}

