% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{single_end_read_table}
\alias{single_end_read_table}
\title{A helper function that creates read tables from single end reads}
\usage{
single_end_read_table(ga, nt_pos)
}
\description{
Should not be called directly by user, but instead is accessed
by calling read_table
}

