% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{single_position.error}
\alias{single_position.error}
\title{Determine the frequency error at variable positions}
\usage{
single_position.error(pu, haplotypes, freq, minimum_coverage = 1000)
}
\arguments{
\item{pu}{A VSeqTools_Pileup object as produced by BAM_pileup}

\item{haplotypes}{character.vector of haplotypes}

\item{freq}{numeric.vector of haplotype frequencies}

\item{min_coverage}{minimum coverage needed to compute errors, otherwise
an NA is returned}
}
\value{
a data.frame containing columns pos, error that give
reference position and frequency error at the position.  Errors are the
sum of absolute value of frequency error over the nucleotides.
}
\description{
Determine the frequency error at variable positions
}
