% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{solution_accuracy.pipeline}
\alias{solution_accuracy.pipeline}
\title{At each position, plot the pileup frequency for a given nucleotide
against the predicted frequency based on the RegressHaplo solution}
\usage{
solution_accuracy.pipeline(
  out_dir,
  solution,
  nucs = c("A", "C", "G", "T"),
  variable_positions = NULL,
  plot = T
)
}
\arguments{
\item{out_dir}{output directory with RegressHaplo files}

\item{solution}{solution number}

\item{variable_positions}{a numeric vector giving the variable positions to 
include.  If NULL, then all variable positions are included (returned
by get_variable_positions.pipeline)}

\item{plot}{should the frequencies be plotted?}

\item{nuc}{which nucleotide shoud be analyzed, one of "A", "C", "G", "T", "d", "i"}
}
\value{
a data.frame giving the true and predicted frequency at each
position, nucleotide combination and a ggplot object of the frequencies
}
\description{
At each position, plot the pileup frequency for a given nucleotide
against the predicted frequency based on the RegressHaplo solution
}
