% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_create_solutions.R
\name{solutions.RegressHaplo}
\alias{solutions.RegressHaplo}
\title{Solve the RegressHaplo optimization repeatedly}
\usage{
solutions.RegressHaplo(
  y,
  P,
  num_trials,
  rho_vals = c(0.1, 1, 2, 5),
  kk_vals = 2
)
}
\value{
A matrix with each column corresponding to a solution.
Within each column, the first four rows are
fit, rho, K, and kk followed the computed solution.
}
\description{
For fixed y and P, generate solutions for every combination of
rho and kk in rho_vals and kk_vals.
}
