% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{split.read_table}
\alias{split.read_table}
\title{Split a read table into two read tables.}
\usage{
\method{split}{read_table}(df)
}
\value{
A list of two read tables labeled df1 and df2.  If the read table
has only 1 position, then NULL is returned.
}
\description{
Split a read table into two read tables.
}

