% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFit.R
\name{split_allele.readFit}
\alias{split_allele.readFit}
\title{Split alleles while accounting for insertions}
\usage{
split_allele.readFit(allele)
}
\arguments{
\item{character}{string}
}
\value{
a character.vector
}
\description{
Split alleles while accounting for insertions
}
\details{
insertions in read tables are stored as, eg., <TTA>.
Alleles for a template formed by pasting, eg, TAT<TTA>CAC.
To recover the read position values we need to split the allele
to return T,A,T,<TTA>,C,A,C
}

