% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{split_paired_ends.read_table}
\alias{split_paired_ends.read_table}
\title{Split reads which have a gap between the paired ends}
\usage{
split_paired_ends.read_table(df)
}
\value{
a data.frame with split reads and a matrix containing
rows that are paired in the returned data.frame
}
\description{
Split reads which have a gap between the paired ends
}
\details{
reads
with no gap are not split, even though they might be from
different ends
}

