% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{split_unlinked_loci.read_table}
\alias{split_unlinked_loci.read_table}
\title{Split a read_table into multiple read_tables representing unlinked loci}
\usage{
split_unlinked_loci.read_table(df, sort_loci = T, min_cover = 1000)
}
\description{
Split a read_table into multiple read_tables representing unlinked loci
}

