% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{subset.read_table}
\alias{subset.read_table}
\title{Limit read table to a subset of positions}
\usage{
\method{subset}{read_table}(df, pos = NULL, start_pos = NULL, end_pos = NULL)
}
\description{
Limit read table to a subset of positions
}
