% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{template_alleles.read_table}
\alias{template_alleles.read_table}
\title{Return all alleles matching a read template}
\usage{
template_alleles.read_table(template, df, match = F)
}
\arguments{
\item{template}{A logical vector}

\item{df}{a read table}

\item{match}{if T then only return alleles that comprise
an entire read.  if F then return all alleles that comprise
any subsequence of a read.}
}
\value{
A named numeric vector with entries giving allele counts
and names giving alleles
}
\description{
Return all alleles matching a read template
}
\details{
a template is a logical vector of length
equal to the number of positions in the read table
and with T corresponding to nucleotide positions that are
part of the template.
}

