% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFit.R
\name{template_fit.readFit}
\alias{template_fit.readFit}
\title{Build a readFit object base on templates.  This is a
generic function used to build readFit objects for
both position and read based fits}
\usage{
template_fit.readFit(template_m, df, h, pi = NULL)
}
\description{
Build a readFit object base on templates.  This is a
generic function used to build readFit objects for
both position and read based fits
}

