% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{templates.read_table}
\alias{templates.read_table}
\title{Retrieve all templates matching at least one read in the
read table}
\usage{
templates.read_table(df)
}
\arguments{
\item{df}{read table}
}
\value{
a logical matrix with each row giving a template
}
\description{
Each read determines a template, with T at the positions
covered by the read, but multiple reads can have the same
template.  This function returns all unique templates.
}
