% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplo.R
\name{unique.Haplo}
\alias{unique.Haplo}
\title{Returns Haplo object with identical haplotype frequencies merged.}
\usage{
\method{unique}{Haplo}(H)
}
\description{
Returns Haplo object with identical haplotype frequencies merged.
}

