% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{unlinked_pos.read_table}
\alias{unlinked_pos.read_table}
\title{Determines loci in read table that have no overlapping reads}
\usage{
unlinked_pos.read_table(df, min_cover)
}
\arguments{
\item{df}{a read table object}
}
\value{
A named logical vector with an entry for every column, other than
count, in the read table data.frame.  Names of entries are the corresponding
column names (positions) in the read tabel data.frame.
}
\description{
Determines loci in read table that have no overlapping reads
}
\details{
If an entry in the returned logical vector is true then all
positions prior to the entry are unlinked to all positions corresponding
to the entry and greater.
}

