% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{variant_calls}
\alias{variant_calls}
\title{Returns variant calls given a VSeqTools_pileup object}
\usage{
variant_calls(pu, min_freq = NULL, sig = NULL, heavy_tail = NULL)
}
\arguments{
\item{pu}{a VSeqTools_pileup object returned by BAM_pileup}

\item{min_freq}{Minimum frequency above which a variant is called.}

\item{sig}{Significance levels above which a variant is called}

\item{heavy_tail}{If T, use betabinomial error distribution, otherwise
use poisson error distribution. Only relavent when !is.null(sig)}
}
\value{
A VSeqTools_variant_call object which inherits the data.frame class.
VSeqTools_variant_call objects are
data.frames with
columns pos, A, C, G, T, d, i.  pos gives the position on
the reference.  Entries in other columns are logical depending
on whether associated variant/deletion/insertion is called
}
\description{
Given a pileup, variants are called accorging to a frequency
cutoff (min_freq) or a significance level(sig)
}
\details{
Exactly one of min_freq and sig must be non-NULL.  This
determines whether cutoff or poisson algorithm is used.
}
