% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{variant_calls_cutoff}
\alias{variant_calls_cutoff}
\title{Returns variant calls given a VSeqTools_pileup object}
\usage{
variant_calls_cutoff(pu, min_freq = 0.03)
}
\arguments{
\item{pu}{a VSeqTools_pileup object returned by BAM_pileup}

\item{min_freq}{Minimum frequency above which a variant is called.}
}
\value{
A VSeqTools_variant_call object which inherits the data.frame class.
VSeqTools_variant_call objects are
data.frames with
columns pos, A, C, G, T, d, i.  pos gives the position on
the reference.  Entries in other columns are logical depending
on whether associated variant/deletion/insertion is called

Warning:  Currently uses the df data.frame in VSeqTools_pileup
object.  Needs to be extended to allow for different strands.
}
\description{
Given a pileup, variants with variability above min_freq are
identified.  This is a simple variant caller which is much faster
than VPhaser2, but cannot call low frequency variants with precision
since phasing information is not used
}
