% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSeqTools_pileup.R
\name{variant_calls_sig}
\alias{variant_calls_sig}
\title{Returns variant calls given a VSeqTools_pileup object}
\usage{
variant_calls_sig(pu, sig = 0.01, heavy_tail = T)
}
\arguments{
\item{pu}{a VSeqTools_pileup object returned by BAM_pileup}

\item{sig}{Significance level of poisson test}

\item{heavy_tail}{If T use betabinomial error distribution, otherwise
use poisson error distribution.}
}
\value{
A VSeqTools_variant_call object which inherits the data.frame class.
VSeqTools_variant_call objects are
data.frames with
columns pos, A, C, G, T, d, i.  pos gives the position on
the reference.  Entries in other columns are logical depending
on whether associated variant/deletion/insertion is called
}
\description{
Given a pileup, implements a poisson significance test to
call variants.
}
\references{
Wang,C. et al. (2007) Characterization of mutation
spectra with ultra-deep pyrosequencing: application to HIV-1 drug
resistance. Genome Res., 17, 1195–1201,

Gerstung et al. (2011) Reliable detection of subclonal
single-nucleotide variants in tumour cell populations.  Nature
Communications.
}
