% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{solutions_to_haplotypes.pipeline}
\alias{solutions_to_haplotypes.pipeline}
\title{Given a RegressHaplo solutions file, determine optimal haplotypes}
\usage{
solutions_to_haplotypes.pipeline(out_dir)
}
\arguments{
\item{out_dir}{output directory for haplotype file and assumed
directory of solutions file}
}
\value{
path to solutions file
}
\description{
Given a RegressHaplo solutions file, determine optimal haplotypes
}

