% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressHaplo_workflow.R
\name{variant_calls_to_read_table.pipeline}
\alias{variant_calls_to_read_table.pipeline}
\title{Given variant calls and a bam file, create a read table}
\usage{
variant_calls_to_read_table.pipeline(bam_file, out_dir, sig = 0.01)
}
\arguments{
\item{bam_file}{path to bam file}

\item{out_dir}{output directory, will be created if needed}

\item{sig}{significance levels at which to filter reads.}
}
\value{
path to read table file
}
\description{
Given variant calls and a bam file, create a read table
}
\details{
Converts bam_file to a raw read table, written to
$outdir/raw_read_table.csv and then applies error correction
to reads and creates a filtered read table, written to
$outdir/read_table.csv
}

